설문에 기간을 정합니다

+++ app/controllers/feedbacks_controller.rb
@@ -4,15 +4,18 @@ class FeedbacksController < ApplicationController
   def create
     option = Option.find params[:option_id]
     survey = option.survey
-    previous_feedback = survey.feedbacks.find_by user: current_user
 
-    if previous_feedback.present?
-      previous_feedback.destroy
-      if previous_feedback.option != option
+    if survey.open?
+      previous_feedback = survey.feedbacks.find_by user: current_user
+
+      if previous_feedback.present?
+        previous_feedback.destroy
+        if previous_feedback.option != option
+          feedback = create_feedback(option)
+        end
+      else
         feedback = create_feedback(option)
       end
-    else
-      feedback = create_feedback(option)
     end
 
     @post = Post.find_by! survey: survey
